+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_03_16_100421) do
+ActiveRecord::Schema.define(version: 2020_03_30_135833) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -142,6 +142,17 @@ ActiveRecord::Schema.define(version: 2020_03_16_100421) do
     t.index ["current_school_id"], name: "index_maths_and_physics_eligibilities_on_current_school_id"
   end
 
+  create_table "notes", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.string "body"
+    t.uuid "claim_id"
+    t.uuid "created_by_id"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["claim_id", "created_at"], name: "index_notes_on_claim_id_and_created_at", unique: true
+    t.index ["claim_id"], name: "index_notes_on_claim_id"
+    t.index ["created_by_id"], name: "index_notes_on_created_by_id"
+  end
+
   create_table "payments", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "payroll_run_id"
     t.decimal "award_amount", precision: 7, scale: 2
@@ -250,6 +261,8 @@ ActiveRecord::Schema.define(version: 2020_03_16_100421) do
   add_foreign_key "claims", "payments"
   add_foreign_key "decisions", "dfe_sign_in_users", column: "created_by_id"
   add_foreign_key "maths_and_physics_eligibilities", "schools", column: "current_school_id"
+  add_foreign_key "notes", "claims"
+  add_foreign_key "notes", "dfe_sign_in_users", column: "created_by_id"
   add_foreign_key "payments", "payroll_runs"
   add_foreign_key "payroll_runs", "dfe_sign_in_users", column: "confirmation_report_uploaded_by_id"
   add_foreign_key "payroll_runs", "dfe_sign_in_users", column: "created_by_id"