+++ spec/support/request_helpers.rb
@@ -22,6 +22,13 @@ module RequestHelpers
     }.fetch(policy)
   end
 
+  def non_service_operator_roles
+    [
+      DfeSignIn::User::PAYROLL_OPERATOR_DFE_SIGN_IN_ROLE_CODE,
+      DfeSignIn::User::SUPPORT_AGENT_DFE_SIGN_IN_ROLE_CODE
+    ]
+  end
+
   # Signs in as a user with the service operator role. Returns the signed-in User record.
   def sign_in_as_service_operator
     user = create(:dfe_signin_user)